requirements.txt
pytest-playwright
Faker
pytest
